refactor: convert app to distributed services with cmd/server binaries and sharing service
